Increased the range for color.

+++ src/main/java/ibis/structure/TouchSet.java
@@ -2,17 +2,17 @@ package ibis.structure;
 
 public final class TouchSet {
   private int numVariables = 0;
-  private byte currentColor = 0;
-  private byte[] colors = null;
+  private short currentColor = 0;
+  private short[] colors = null;
 
   public TouchSet(int numVariables) {
     this.numVariables = numVariables;
-    this.colors = new byte[2 * numVariables + 1];
+    this.colors = new short[2 * numVariables + 1];
     reset();
   }
 
   public void reset() {
-    if (currentColor == Byte.MAX_VALUE) {
+    if (currentColor == Short.MAX_VALUE) {
       java.util.Arrays.fill(colors, (byte) 0);
       currentColor = 0;
     }
